cardsignalhq_repo_regenerated/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.2
fastapi==0.115.0
uvicorn[standard]==0.30.6
